backend/pyproject.toml
[project]
name = "running-manager-backend"
version = "0.1.0"
description = "Running Manager API (FastAPI)"
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "uvicorn",
    "sqlalchemy",
    "passlib",
    "bcrypt==4.0.1",
    "python-jose[cryptography]",
    "python-multipart",
    "lxml",
    "python-dateutil",
    "aiofiles",
    "pymysql",
    "python-dotenv>=1.2.1",
    "langchain>=0.3",
    "langchain-anthropic>=0.3",
    "langchain-openai>=0.3",
    "langgraph>=0.2",
    "langfuse>=2",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "httpx>=0.28",
    "freezegun>=1.4",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
pythonpath = ["."]
